Add files via upload
This is my first try of using vivado to program the fpga to send a pulse to trigger the camera. Based on the LED_Blink tutorial from http://antonpotocnik.com/?p=487360 I simply changed the bit that is used to turn on the led from the 26th to the 27th (Longer pulses basically). Finally, I routed the output to the "exp_p_tri_io" port to use the digital 3,3 V pin on the red pitaya. The pin goes high each time the LED turns on.