Add files via upload
In this version, we can use the linux terminal inside the RP to trigger the camera whenever the user wants. It is not hooked up to a clock or anything, it just waits for the user to issue the command to raise or lower the voltage. Based on https://antonpotocnik.com/?p=489265 stopwatch tutorial.