Final C contribution of Carlos' part
FlyCapture2Test_C is an EXAMPLE file of how to use the API, it is not actually used in the program itself.  Use it for reference.
RPserver.c is copied and compiled inside the RP's terminal through putty.
main.c is the main file obviously
Functions.c has all the definitions of the functions, which are declared in the header Functionns.h
FlyCapture2Platform_C, FlyCapture2_C and Flycapture2Defs are the only headers from the FlyCapture2 api necessary for the program.
The FlyCapture2_C_v100.lib file has to be linked in order for the program to run 
Linking the ws2_32 library is necessary for the sockets in the client server connection to run. 
I'm including my cameraCCode.cbp file for codeblocks IDE. Go into build options -> Linker settings to see how to link the libraries